Correct [WindowsHotkeyRegistrar] behaviour.
The WinAPI [RegisterHotKey] function does not allow the registering of
two hotkeys with the same combination and the same Window handle value
if the provided IDs differ. As such, there is no way for the registrar
to differentiate between one hotkey registered with the no-repeat flag
and one registered without (where they use the same combination).

The registrar now fails on attempts to register such hotkeys.