Distinguish tests applying to whole object.
Tests are named for the portion of the device under test they relate to,
causing the tests to be grouped in reviewing tools (with an alphabetical
sort). However, including the class name in tests which apply to a class
generally (rather than a particular unit or function) is redundant.

Such tests are now prefixed with an underscore to cause their grouping.